+++ CHANGELOG.md
@@ -20,5 +20,6 @@ Released changes are shown in the
 ### Removed
 
 ### Fixed
+- Fixed utterances table poorly showing ids greater than 9999.
 
 ### Security